storable: add `StorableVec`, remove `Storable::ALIGN`
